+++ models/stock_movement.go
@@ -18,7 +18,7 @@ type StockMovement struct {
 	TenantID    uint                `json:"tenant_id" gorm:"index;not null"`
 	ProductID   uint                `json:"product_id" gorm:"index;not null"`
 	ProductName string              `json:"product_name" gorm:"size:200"`
-	Delta       int                 `json:"delta"`      // positive = load, negative = deduction
+	Delta       float64             `json:"delta"`      // positive = load, negative = deduction
 	Reason      StockMovementReason `json:"reason" gorm:"type:varchar(30);not null"`
 	ReferenceID *uint               `json:"reference_id,omitempty"` // sale_id when reason=sale
 	Notes       string              `json:"notes" gorm:"size:300"`
